+++ kernel/packages/unity-interface/BrowserInterface.ts
@@ -21,12 +21,7 @@ import {
   WorldPosition,
   LoadableParcelScene
 } from 'shared/types'
-import {
-  forceStopParcelSceneWorker,
-  getSceneWorkerBySceneID,
-  setNewParcelScene,
-  stopParcelSceneWorker
-} from 'shared/world/parcelSceneManager'
+import { getSceneWorkerBySceneID, setNewParcelScene, stopParcelSceneWorker } from 'shared/world/parcelSceneManager'
 import { getPerformanceInfo, getRawPerformanceInfo } from 'shared/session/getPerformanceInfo'
 import { positionObservable } from 'shared/world/positionThings'
 import { renderStateObservable } from 'shared/world/worldState'
@@ -55,6 +50,7 @@ import { ensureFriendProfile } from 'shared/friends/ensureFriendProfile'
 import Html from 'shared/Html'
 import { reloadScene } from 'decentraland-loader/lifecycle/utils/reloadScene'
 import { isGuest } from '../shared/ethereum/provider'
+import { killPortableExperienceScene } from './portableExperiencesUtils'
 
 declare const DCL: any
 
@@ -469,13 +465,8 @@ export class BrowserInterface {
     }
   }
 
-  public KillPortableExperience(data: { portableExperienceId: string }) {
-    const peWorker = getSceneWorkerBySceneID(data.portableExperienceId)
-
-    if (peWorker) {
-      forceStopParcelSceneWorker(peWorker)
-    }
-    unityInterface.UnloadScene(data.portableExperienceId)
+  public async KillPortableExperience(data: { portableExperienceId: string }): Promise<void> {
+    await killPortableExperienceScene(data.portableExperienceId)
   }
 }
